issue 46
Added functionality to watch the movies and images directory for changes using JNotify.
Be aware, that you will need the "lib" folder in java.library.path (like -Djava.library.path="lib")
The updates are deferred, so a lot of fast changes in the directories shouldnt be a problem.

Please test on Linux and Windows



git-svn-id: http://blinkendroid.googlecode.com/svn/branches/development@636 ded441ff-28c8-dbd9-35dd-4b5d9abe79e0